# Conflicts:
#	target/classes/repository/ClientRepository.class
#	target/classes/repository/FreelancerRepository.class
#	target/classes/security/SecurityConfig.class

+++ src/main/java/dto/AddReviewRequestDto.java
@@ -1,10 +1,12 @@
 package dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class AddReviewRequestDto {
     public String revieweeId;
-    public double rating;
+    public BigDecimal rating;
     public String details;
 }
